requirements.txt
pandas
requests
pyodbc
python-dotenv
sqlalchemy
psycopg2